Add middleware to log errors
This proposes the split between logging and converting errors into
response.

The idea is to always add a debug message with the error that happened,
which is handy for both production and development.
In combination with an access log middleware we can log errors or
warnings, depending on the final response status code.

For production, we assume that a fingers crossed handler is used so that
only errors are sent to log - with enough information.

More info:

 - https://github.com/middlewares/access-log/blob/master/src/AccessLog.php
 - https://github.com/Seldaek/monolog/blob/master/src/Monolog/Handler/FingersCrossedHandler.php